level 3 (dernier)
ajout du 3e niveau : un corridor avec 4 chevalier et 2 golems (il est vraiment plus dur mais c'est pas grave)

+++ Assets/Arthur/Scripts/Enemy.cs
@@ -3,11 +3,11 @@ using UnityEngine;
 
 public class Enemy : MonoBehaviour
 {
-    public int speed = 10;
-    public int baseSpeed;
+    public float speed = 10;
+    public float baseSpeed;
     //public EnemyStatus status;
     public Path path;
-    public Enemy(int speed, Path path)
+    public Enemy(float speed, Path path)
     {
         this.speed = speed;
         this.path = path;
